Move code to a new module
